Modified fromSecondsUsingDays().
Changed the docblock, made the method static, adding additional "days" value check, and throwing an exception if it is not set. The method was made static since "days" can not be set by anything aside from "DateTime::diff()". The additional check for "false" was added since the documentation claims that "false" is returned if it is not set by a diff().